refactor(cli): the verb arms fold out of main.rs under the file wall
main.rs crossed 1500 lines with the two case flags on the test verb.
The check, test and inspect arms move to a child module (arms.rs);
the dispatch seam stays one line per verb.
